+++ src/main/groovy/nextflow/validation/help/HelpObserver.groovy
@@ -3,7 +3,7 @@ package nextflow.validation.help
 import groovy.util.logging.Slf4j
 
 import nextflow.processor.TaskHandler
-import nextflow.trace.TraceObserver
+import nextflow.trace.TraceObserverV2
 import nextflow.trace.TraceRecord
 import nextflow.Session
 
@@ -11,13 +11,13 @@ import nextflow.validation.help.HelpMessageCreator
 import nextflow.validation.config.ValidationConfig
 
 @Slf4j
-class HelpObserver implements TraceObserver {
+class HelpObserver implements TraceObserverV2 {
     
     @Override
     void onFlowCreate(Session session) {
         // Help message logic
         def Map params = (Map)session.params ?: [:]
-        def ValidationConfig config = new ValidationConfig(session?.config?.navigate('validation') as Map, params)
+        def ValidationConfig config = new ValidationConfig(session?.config?.navigate('validation') as Map, session)
         def Boolean containsFullParameter = params.containsKey(config.help.fullParameter) && params[config.help.fullParameter]
         def Boolean containsShortParameter = params.containsKey(config.help.shortParameter) && params[config.help.shortParameter]
         if (config.help.enabled && (containsFullParameter || containsShortParameter)) {