+++ src/test/groovy/nextflow/validation/ValidateParametersTest.groovy
@@ -246,7 +246,7 @@ class ValidateParametersTest extends Dsl2Spec{
         then:
         def error = thrown(SchemaValidationException)
         def errorMessages = error.message.readLines()
-        errorMessages[0] == "\033[0;31mThe following invalid input values have been detected:"
+        errorMessages[0] == "The following invalid input values have been detected:"
         errorMessages[1] == ""
         errorMessages[2] == "* --input (src/testResources/wrong.csv): Validation of file failed:"
         errorMessages[3] == "\t-> Entry 1: Error for field 'strandedness' (weird): Expected any of [[forward, reverse, unstranded]] (Strandedness must be provided and be one of 'forward', 'reverse' or 'unstranded')"
@@ -280,7 +280,7 @@ class ValidateParametersTest extends Dsl2Spec{
         then:
         def error = thrown(SchemaValidationException)
         def errorMessages = error.message.readLines()
-        errorMessages[0] == "\033[0;31mThe following invalid input values have been detected:"
+        errorMessages[0] == "The following invalid input values have been detected:"
         errorMessages[1] == ""
         errorMessages[2] == "* --input (src/testResources/wrong.tsv): Validation of file failed:"
         errorMessages[3] == "\t-> Entry 1: Error for field 'strandedness' (weird): Expected any of [[forward, reverse, unstranded]] (Strandedness must be provided and be one of 'forward', 'reverse' or 'unstranded')"
@@ -314,7 +314,7 @@ class ValidateParametersTest extends Dsl2Spec{
         then:
         def error = thrown(SchemaValidationException)
         def errorMessages = error.message.readLines()
-        errorMessages[0] == "\033[0;31mThe following invalid input values have been detected:"
+        errorMessages[0] == "The following invalid input values have been detected:"
         errorMessages[1] == ""
         errorMessages[2] == "* --input (src/testResources/wrong.yaml): Validation of file failed:"
         errorMessages[3] == "\t-> Entry 1: Error for field 'strandedness' (weird): Expected any of [[forward, reverse, unstranded]] (Strandedness must be provided and be one of 'forward', 'reverse' or 'unstranded')"
@@ -348,7 +348,7 @@ class ValidateParametersTest extends Dsl2Spec{
         then:
         def error = thrown(SchemaValidationException)
         def errorMessages = error.message.readLines()
-        errorMessages[0] == "\033[0;31mThe following invalid input values have been detected:"
+        errorMessages[0] == "The following invalid input values have been detected:"
         errorMessages[1] == ""
         errorMessages[2] == "* --input (src/testResources/wrong.json): Validation of file failed:"
         errorMessages[3] == "\t-> Entry 1: Error for field 'strandedness' (weird): Expected any of [[forward, reverse, unstranded]] (Strandedness must be provided and be one of 'forward', 'reverse' or 'unstranded')"
@@ -611,8 +611,10 @@ class ValidateParametersTest extends Dsl2Spec{
 
         when:
         def config = ["validation": [
-            "failUnrecognisedParams": true,
-            "monochromeLogs": true
+            "monochromeLogs": true,
+            "logging": [
+                "unrecognisedParams": "error"
+            ]
         ]]
         def result = new MockScriptRunner(config).setScript(SCRIPT).execute()
         def stdout = capture
@@ -795,75 +797,6 @@ class ValidateParametersTest extends Dsl2Spec{
         def schema = Path.of('src/testResources/nextflow_schema_file_path_pattern.json').toAbsolutePath().toString()
         def SCRIPT = """
             params.glob = 'src/testResources/correct.csv'
-            include { validateParameters } from 'plugin/nf-schema'
-            
-            validateParameters(parameters_schema: '$schema')
-        """
-
-        when:
-        def config = [:]
-        def result = new MockScriptRunner(config).setScript(SCRIPT).execute()
-        def stdout = capture
-                .toString()
-                .readLines()
-                .findResults {it.contains('WARN nextflow.validation.SchemaValidator') || it.startsWith('* --') ? it : null }
-
-        then:
-        noExceptionThrown()
-        !stdout
-    }
-
-    def 'ignore validation of aws s3' () {
-        given:
-        def schema = Path.of('src/testResources/nextflow_schema_file_path_pattern.json').toAbsolutePath().toString()
-        def SCRIPT = """
-            params.glob = 's3://src/testResources/correct.csv'
-            include { validateParameters } from 'plugin/nf-schema'
-            
-            validateParameters(parameters_schema: '$schema')
-        """
-
-        when:
-        def config = [:]
-        def result = new MockScriptRunner(config).setScript(SCRIPT).execute()
-        def stdout = capture
-                .toString()
-                .readLines()
-                .findResults {it.contains('WARN nextflow.validation.SchemaValidator') || it.startsWith('* --') ? it : null }
-
-        then:
-        noExceptionThrown()
-        !stdout
-    }
-
-    def 'ignore validation of azure blob storage' () {
-        given:
-        def schema = Path.of('src/testResources/nextflow_schema_file_path_pattern.json').toAbsolutePath().toString()
-        def SCRIPT = """
-            params.glob = 'az://src/testResources/correct.csv'
-            include { validateParameters } from 'plugin/nf-schema'
-            
-            validateParameters(parameters_schema: '$schema')
-        """
-
-        when:
-        def config = [:]
-        def result = new MockScriptRunner(config).setScript(SCRIPT).execute()
-        def stdout = capture
-                .toString()
-                .readLines()
-                .findResults {it.contains('WARN nextflow.validation.SchemaValidator') || it.startsWith('* --') ? it : null }
-
-        then:
-        noExceptionThrown()
-        !stdout
-    }
-
-    def 'ignore validation of gcp cloud storage' () {
-        given:
-        def schema = Path.of('src/testResources/nextflow_schema_file_path_pattern.json').toAbsolutePath().toString()
-        def SCRIPT = """
-            params.glob = 'gs://src/testResources/correct.csv'
             include { validateParameters } from 'plugin/nf-schema'
             
             validateParameters(parameters_schema: '$schema')